fix(network): redact credentials from registry-retry log lines
A registry can be configured with inline basic-auth
(`https://user:pass@host/`), which pacquet supports. The retry warnings
in this module logged the request URL verbatim, leaking those
credentials into CI logs and log aggregators — exactly when logs are
captured for diagnosis. The user-facing fetch error is already redacted
by the verifier, but the new `retry_async` log and the existing
`send_with_retry` logs were not.

Hoist the `redact_url_credentials` helper (a port of pnpm's
`redactUrlCredentials`) from the npm-resolver's verifier — where it was a
private copy — into `pacquet_network::auth`, export it, and apply it at
all three retry log sites. The verifier now imports the shared helper
instead of carrying its own, removing the duplication.
